Rename namespace NuSoftHep nusofthep. Migrate package.py
