+++ spack_repo/nusofthep/packages/dk2nudata/package.py
@@ -7,7 +7,8 @@ import os
 
 from spack import *
 from spack.package import *
-from spack.pkg.fnal_art.fnal_github_package import *
+from spack_repo.builtin.build_systems.cmake import CMakePackage
+from spack_repo.fnal_art.packages.fnal_github_package.package import *
 
 
 class Dk2nudata(CMakePackage, FnalGithubPackage):
