+++ spack_repo/nusofthep/packages/nuevdb/package.py
@@ -5,20 +5,19 @@
 
 from spack import *
 from spack.package import *
-from spack.pkg.fnal_art.fnal_github_package import *
+from spack_repo.builtin.build_systems.cmake import CMakePackage
+from spack_repo.fnal_art.packages.fnal_github_package.package import *
 
 
-class Nutools(CMakePackage, FnalGithubPackage):
-    """Nutools"""
+class Nuevdb(CMakePackage, FnalGithubPackage):
+    """Nuevdb"""
 
-    repo = "NuSoftHEP/nutools"
+    repo = "NuSoftHEP/nuevdb"
     license("Apache-2.0")
-    version_patterns = ["v3_15_04", "3.16.03"]
+    version_patterns = ["v1_08_01", "1.09.03"]
 
-    version("3.17.00", sha256="48b6be64291411d27878da5010415564fd658ecffc72e617fb8a11579ecaed0b")
-    version("3.16.06", sha256="f540be7b30eec357c5f65260be6da3ce6988e5b193c58770baaa36a913a513ac")
-                
     version("develop", branch="develop", get_full_repo=True)
+    version("1.09.08", sha256="bc949e57ecc9a1658606decd884dc736fa130539d746211231d9b9a60a18a745")
 
     cxxstd_variant("17", "20", default="17")
 
@@ -29,17 +28,26 @@ class Nutools(CMakePackage, FnalGithubPackage):
 
     depends_on("art")
     depends_on("art-root-io")
-    depends_on("cry")
+    depends_on("boost+date_time")
+    depends_on("cetlib")
+    depends_on("cetlib-except")
+    depends_on("libwda")
     depends_on("nusimdata")
-    depends_on("perl")
+    depends_on("postgresql")
+    depends_on("root")
+
+    with when("@:1.09.07"):
+        depends_on("canvas-root-io")
+        depends_on("boost +filesystem +regex +thread")
 
     @cmake_preset
     def cmake_args(self):
         return [
             self.define_from_variant("CMAKE_CXX_STANDARD", "cxxstd"),
             self.define("IGNORE_ABSOLUTE_TRANSITIVE_DEPENDENCIES", True),
+            self.define("libwda_DIR:PATH", self.spec["libwda"].prefix),
         ]
 
     @sanitize_paths
     def setup_run_environment(self, run_env):
-        run_env.prepend_path("ROOT_INCLUDE_PATH", self.prefix.include)
+        run_env.prepend_path("CET_PLUGIN_PATH", self.prefix.lib)
