+++ spack_repo/nusofthep/packages/nurandom/package.py
@@ -5,7 +5,8 @@
 
 from spack import *
 from spack.package import *
-from spack.pkg.fnal_art.fnal_github_package import *
+from spack_repo.builtin.build_systems.cmake import CMakePackage
+from spack_repo.fnal_art.packages.fnal_github_package.package import *
 from spack.util.prefix import Prefix
