+++ spack_repo/nusofthep/repo.yaml
@@ -1,0 +1,3 @@
+repo:
+  namespace: nusofthep
+  api: v2.0
